module7-task1: use AMD for review form
