+++ src/main.js
@@ -1,5 +1,4 @@
 'use strict';
 
-require('./form');
-require('./game');
-require('./reviews');
+require(['./form/form']);
+
